==> notebooks/1-discover.ipynb <==
# 1. Discover 
#### A deep dive into the problem we are trying to solve

> Successful people have fear, successful people have doubts, and successful people have worry. They just don’t let these feelings stop them.

This section covers the discover phase of the project lifecycle

This phase covers:
- An idea of how AI can help solve a particular problem (done)
    - Inspirtation of this idea (done)
- Building research questions (done)
- Literature review of work done in the area of energy (done)
- What is the end in mind? (done)
- Hypothesis formulation (done)


---

#### 1. Ideation

The goal of this thesis is to model energy usage for households using generally
accessible household data, and to highlight the uses for this model, not just for
the home owners, but also general sustainability bodies in government, in the
focus on driving to a net zero target by 2050.

#### 2. Research Questions

**RQ1** 
- What are the data requirements for modelling energy usage for house-
holds?

**RQ2** 
- How should machine learning be applied to best predict household energy
efficiency?

**RQ3** 
- How can users effectively interact with and gain insight about the house-
hold base from this model?

#### 3. Literature Review

Covered in Chapter 3 of Thesis Report

#### 4. End in Mind

The end goal of this work is to develop a model that can predict energy efficiency for home owners and governing bodies, overlay a dashboard GUI structure on top of this that can be used either in order to understand the personalised journey of each home they dealt with.

#### 5. Hypothesis Formulation

**The BER of households can be predicted using general household
characteristics to an accuracy of 80%**
