==> notebooks/3-design.ipynb <==
import pandas as pd
import numpy as np
import pandas as pd
from sklearn.experimental import enable_iterative_imputer
from sklearn.impute import IterativeImputer
from sklearn.ensemble import RandomForestRegressor, RandomForestClassifier
import random
from sklearn.tree import DecisionTreeClassifier
from sklearn.model_selection import RepeatedStratifiedKFold
from sklearn.svm import LinearSVC
from sklearn.naive_bayes import GaussianNB
from imblearn.pipeline import Pipeline
from sklearn.neighbors import KNeighborsClassifier
from sklearn.model_selection import GridSearchCV
from imblearn.over_sampling import SMOTE
from sklearn.linear_model import LogisticRegression
from sklearn.linear_model import LogisticRegressionCV
from sklearn.svm import SVC
from sklearn.preprocessing import MinMaxScaler, StandardScaler, QuantileTransformer, RobustScaler
from imblearn.combine import SMOTETomek
from sklearn.model_selection import train_test_split, cross_validate
from imblearn.under_sampling import TomekLinks
from sklearn.metrics import classification_report
import matplotlib.pyplot as plt
from sklearn.model_selection import cross_val_score
from yellowbrick.classifier.rocauc import roc_auc
from sklearn.preprocessing import OneHotEncoder
from sklearn.compose import make_column_transformer
from sklearn.compose import make_column_selector as selector
from category_encoders import CatBoostEncoder
from yellowbrick.classifier import ROCAUC
from imblearn.over_sampling import SMOTENC
from sklearn.metrics import roc_curve, auc
from imblearn.under_sampling import RandomUnderSampler
import category_encoders as ce
from sklearn.impute import SimpleImputer
from imblearn.over_sampling import RandomOverSampler
from imblearn.pipeline import Pipeline
from sklearn.preprocessing import LabelEncoder
import warnings
warnings.filterwarnings("ignore")
random.seed(2814)

# Read in Imputed dataset
df = pd.read_csv('../data/processed/Define/seai_missingforest_data_imputed.csv')

# Feature Engineering
conditions = [
    (df['MainSpaceHeatingFuel'] == 'Heating Oil'),
    (df['MainSpaceHeatingFuel'] == 'Mains Gas'),
    (df['MainSpaceHeatingFuel'] == 'Electricity'),
    (df['MainSpaceHeatingFuel'] == 'Solid Multi-Fuel'),
    (df['MainSpaceHeatingFuel'] == 'Sod Peat'),
    (df['MainSpaceHeatingFuel'] == 'Bulk LPG (propane or butane)'),
    (df['MainSpaceHeatingFuel'] == 'House Coal'),
    (df['MainSpaceHeatingFuel'] == 'Wood Logs'),
    (df['MainSpaceHeatingFuel'] == 'Bottled LPG'),
    (df['MainSpaceHeatingFuel'] == 'Peat Briquettes'),
    (df['MainSpaceHeatingFuel'] == 'Wood Pellets (bulk supply for'),
    (df['MainSpaceHeatingFuel'] == 'Electricity - Standard Domesti'),
    (df['MainSpaceHeatingFuel'] == 'Wood Pellets (in bags for seco'),
    (df['MainSpaceHeatingFuel'] == 'Manufactured Smokeless Fuel'),
    (df['MainSpaceHeatingFuel'] == 'Anthracite'),
    (df['MainSpaceHeatingFuel'] == 'Wood Chips'),
    (df['MainSpaceHeatingFuel'] == 'Electricity - Off-peak Night-R'),
    (df['MainSpaceHeatingFuel'] == 'Electricity - On-peak Night-Ra'),
    (df['MainSpaceHeatingFuel'] == 'Bioethanol from renewable sour'),
    (df['MainSpaceHeatingFuel'] == 'Biodiesel from renewable sourc')
    ]

# create a list of the values we want to assign for each condition
values = [15.57, 10.00, 7.93, 9.51, 9.51, 15.00, 8.75, 13.50, 24.21, 9.51, 9.07, 7.93, 9.52, 8.80, 8.40, 5.92, 11.54, 11.54, 12.50, 12.50]

# create a new column and use np.select to assign values to it using our lists as arguments
df['EnergyCost'] = np.select(conditions, values, 12.50)
df['EnergyCost'] = round((df['TotalDeliveredEnergy'] * df['EnergyCost']) / 100, 2)
# display updated DataFrame
df.head()

# Move BER to the very end 
new_cols = [col for col in df.columns if col != 'EnergyRating'] + ['EnergyRating']
df = df[new_cols]
del(new_cols)

# Make sure year is not counted as a numerical feature
df['YearofConstruction'] = df['YearofConstruction'].astype('object')

df.EnergyRating.value_counts().sort_index()

##### Outline of steps for data to go through in a pipe:

- Data Ingestion
- Data Cleaning
- Data Imputation
- Data Sampling
- Data Splitting of Categorical and Numerical pipes within main pipe
- Run MinMaxScaler on Numerical data
- Run OneHotEncoding on the 

Now that impute is complete, we need to sample the dataset

# Separate feature and target variables
X = df.iloc[:, :-1] # Independent Variables
y = df.iloc[:, -1] # Dependent Variables

# Randomly distribute the data so CatBoost can work
perm = np.random.permutation(len(X))
X = X.iloc[perm].reset_index(drop=True) 
y = y.iloc[perm].reset_index(drop=True)

# Train-Test Split
X_train, X_test, y_train, y_test = train_test_split(
    X, y, test_size=0.2, stratify=y, random_state=2814)

X_test

# Label Encoding of Target for CatBoost
yle = LabelEncoder()
y_train = yle.fit_transform(y_train)
y_test = yle.transform(y_test)

from imblearn.pipeline import Pipeline as Pipe
pipe = Pipe([('scaler1', CatBoostEncoder()), ('scaler2', MinMaxScaler()), ('smote', SMOTETomek()), ('rf', RandomForestClassifier(random_state=2814, criterion='entropy'))])


grid_search = GridSearchCV(estimator=pipe,
                           param_grid=param_grid,
                           scoring='roc_auc',
                           cv=stratified_kfold,
                           n_jobs=-1)

grid_search.fit(X_train, y_train)
cv_score = grid_search.best_score_
test_score = grid_search.score(X_test, y_test)
print(f'Cross-validation score: {cv_score}\nTest score: {test_score}')

# CatBoost Encoding
cbe_encoder = ce.cat_boost.CatBoostEncoder()
X_train = cbe_encoder.fit_transform(X_train, y_train)
X_test = cbe_encoder.transform(X_test)

df.iloc[209132]

# Taking a look at the encoding
pd.DataFrame(X_train)

# Scaling the dataset. Not required for tree models but good practice.
scaler = MinMaxScaler()
X_train = scaler.fit_transform(X_train, y_train)
X_test = scaler.transform(X_test)

# Use value counts as reference for under sampler
pd.DataFrame(y_train).value_counts().sort_index()

resample=SMOTETomek(random_state=2814)
X_train, y_train = resample.fit_resample(X_train, y_train)

model = RandomForestClassifier(random_state=2814, criterion='entropy', n_estimators=175)
model.fit(X_train, y_train)
y_pred = model.predict(X_test)
print(classification_report(y_test,y_pred))

from sklearn.metrics import confusion_matrix
matrix = confusion_matrix(y_test, y_pred)
matrix.diagonal()/matrix.sum(axis=1)

from numpy import save
save('../data/interim/Design/X_train.npy', X_train)
save('../data/interim/Design/X_test.npy', X_test)
save('../data/interim/Design/y_train.npy', y_train)
save('../data/interim/Design/y_test.npy', y_test)

# Save our final model
import pickle
filename = '../models/final_model.pkl'
pickle.dump(model, open(filename, 'wb'))

import xgboost as xgb
xgb_model = xgb.XGBClassifier(objective="multi:softprob", random_state=2814)
xgb_model.fit(X_train, y_train)
y_pred = xgb_model.predict(X_test)
print(classification_report(y_test, y_pred))
