==> notebooks/4.1-deliver_NN.ipynb <==
import numpy as np # linear algebra
import pandas as pd # data processing, CSV file I/O (e.g. pd.read_csv)
from keras.models import Sequential
from keras.layers import Dense
from keras.layers import Dropout
from sklearn.preprocessing import LabelEncoder
from tensorflow import keras
from scikeras.wrappers import KerasClassifier
import warnings
warnings.filterwarnings("ignore")
import pandas as pd
import numpy as np
import pandas as pd
from sklearn.ensemble import RandomForestClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.model_selection import RepeatedStratifiedKFold
from sklearn.svm import LinearSVC
from sklearn.naive_bayes import GaussianNB
from sklearn.neighbors import KNeighborsClassifier
from sklearn.model_selection import GridSearchCV
from sklearn.linear_model import LogisticRegression
from sklearn.linear_model import LogisticRegressionCV
from sklearn.svm import SVC
from sklearn.model_selection import train_test_split, cross_validate
from sklearn.metrics import classification_report
from sklearn.model_selection import cross_val_score
from sklearn.metrics import roc_curve, auc
from sklearn.neighbors import KNeighborsClassifier
from tensorflow.keras.layers import LeakyReLU, Dense, BatchNormalization
import random


from numpy import load
X_train = load('../data/interim//Design/X_train.npy')
X_test = load('../data/interim/Design/X_test.npy')
y_train = load('../data/interim/Design/y_train.npy')
y_test = load('../data/interim/Design/y_test.npy')

yle = LabelEncoder()
y_train = yle.fit_transform(y_train)
y_test = yle.transform(y_test)
from tensorflow.keras.utils import to_categorical
y_train = to_categorical(y_train, 15)
y_test = to_categorical(y_test, 15)

from tensorflow.keras.layers import Dense

def baseline_model():
    # Create model here
    model = Sequential()
    model.add(Dense(1024, input_dim = 13, activation = 'relu')) # Rectified Linear Unit Activation Function
    model.add(Dense(512, activation = 'relu'))
    model.add(Dense(256, activation = 'relu'))
    model.add(Dense(128, activation = 'relu'))
    model.add(Dense(64, activation = 'relu'))
    model.add(Dense(32, activation = 'relu'))
    model.add(Dense(15, activation = 'softmax')) # Softmax for multi-class classification
    # Compile model here
    model.compile(loss='categorical_crossentropy', optimizer='adam', metrics=['accuracy', keras.metrics.Precision(), keras.metrics.Recall()])
    return model

estimator = KerasClassifier(model=baseline_model, epochs=50, batch_size=128)
estimator.fit(X_train, y_train, validation_data = ((X_test, y_test)))

y_pred = estimator.predict(X_test)
print(classification_report(y_test, y_pred))
